% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footnote.R
\name{add_footnote}
\alias{add_footnote}
\title{Add a footnote to a Typst table}
\usage{
add_footnote(x, content, type = "general", location)
}
\arguments{
\item{x}{A Typst table}

\item{content}{The content of the footnote}

\item{type}{The labelling style for the footnote. Either \code{"general"} (no
label), \code{"number"} (numeric labelling), \code{"alphabet"} (alphabetic
numbering), or \code{"symbol"} (symbolic numbering).}

\item{location}{The cell(s) in which the footnote reference should be
inserted. Should be a \code{cells_location} object created by \code{\link[=cells]{cells()}}. Only
if \code{type != "general"}.}
}
\value{
A Typst table with the footnote added.
}
\description{
Add a footnote to a Typst table
}
