% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-cells.R
\name{add_indent}
\alias{add_indent}
\title{Add indentation to Typst table cells}
\usage{
add_indent(x, location, amount = "1em")
}
\arguments{
\item{x}{A Typst table}

\item{location}{Cells to apply indenting to. Should be a \code{cells_location}
object created by \code{\link[=cells]{cells()}}.}

\item{amount}{Amount of indentation to add. Should be a Typst
\href{https://typst.app/docs/reference/text/text/#parameters-size}{length}
specification. A numeric \code{amount} will be taken as 'em' units.}
}
\value{
A Typst table with indentation added to the specified cells.
}
\description{
Add indentation to Typst table cells
}
