% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_docx.R
\name{as_docx}
\alias{as_docx}
\title{Export a Typst table or figure object as raw Word XML}
\usage{
as_docx(x)
}
\arguments{
\item{x}{A Typst table or figure (\code{ttables_tbl} or \code{ttables_fig})}
}
\value{
A \code{glue} string containing the raw XML output.
}
\description{
Export a Typst table or figure object as raw Word XML
}
