% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-cells.R
\name{format_cells}
\alias{format_cells}
\title{Add formatting to Typst table cells}
\usage{
format_cells(
  x,
  location,
  bold = NULL,
  italic = NULL,
  align = NULL,
  indent = NULL,
  size = NULL,
  stroke = NULL,
  fill = NULL
)
}
\arguments{
\item{x}{A Typst table}

\item{location}{Cells to apply the formating to. Should be a \code{cells_location}
object created by \code{\link[=cells]{cells()}}.}

\item{bold, italic}{Whether to apply bold or italic styling to the cell(s).}

\item{align}{How to align the cells contents. Either \code{"auto"} or a Typst
\href{https://typst.app/docs/reference/layout/alignment/}{alignment}
specification.}

\item{indent}{How much to indent cell contents. Should be a Typst
\href{https://typst.app/docs/reference/layout/length/}{length} specification.
A numeric \code{indent} will be taken as 'em' units.}

\item{size}{Font size. Should be a Typst
\href{https://typst.app/docs/reference/layout/length/}{length} specification.
A numeric \code{size} will be taken as points.}

\item{stroke}{Cell stroke (border) style. Should be Typst
\href{https://typst.app/docs/reference/visualize/stroke/}{stroke}
specification. Only the paint and thickness components are supported for
now.}

\item{fill}{Cell fill colour. Should be Typst
\href{https://typst.app/docs/reference/visualize/color/}{color}
specification.}
}
\value{
A Typst table with formatting specification applied to the given
cell(s).
}
\description{
Add formatting to Typst table cells
}
