% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{landscape}
\alias{landscape}
\title{Put a Typst table on its own landscape page}
\usage{
landscape(x)
}
\arguments{
\item{x}{A Typst table}
}
\value{
A Typst table with the landscape page specification.
}
\description{
Put a Typst table on its own landscape page
}
