% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_rows.R
\name{pack_rows}
\alias{pack_rows}
\title{Group table rows together under a group label}
\usage{
pack_rows(
  x,
  start,
  nrows,
  label = NULL,
  italic = TRUE,
  bold = FALSE,
  align = NULL,
  indent = TRUE
)
}
\arguments{
\item{x}{A Typst table}

\item{start}{The first row of the group}

\item{nrows}{The number of rows to group together}

\item{label}{The label to add above the grouped rows. \code{NULL} for no label.}

\item{italic, bold}{Whether the group label should be formatted with italic
and/or bold styling. Default is italic, regular weight.}

\item{align}{Cell alignment for the group label. Either \code{"auto"} or a Typst
\href{https://typst.app/docs/reference/layout/alignment/}{alignment}
specification. \code{NULL} (the default) to use the table alignment.}

\item{indent}{Whether to indent the grouped rows. Either \code{TRUE} for the
default (\verb{1em}) indent, \code{FALSE} for no indent, or a Typst
\href{https://typst.app/docs/reference/text/text/#parameters-size}{length}
specification. A numeric \code{indent} will be taken as 'em' units. Default is
\code{TRUE.}}
}
\value{
A Typst table with the specified grouped rows.
}
\description{
Group table rows together under a group label
}
