% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_figure_options.R
\name{set_figure_options}
\alias{set_figure_options}
\title{Set Typst figre options}
\usage{
set_figure_options(
  x,
  width,
  height,
  placement,
  caption,
  label,
  supplement,
  landscape
)
}
\arguments{
\item{x}{A Typst figure}

\item{width, height}{Figure width and height. Either \code{"auto"} or a Typst
\href{https://typst.app/docs/reference/visualize/image/#parameters-width}{length}
specification.}

\item{placement}{Figure placement. Either \code{"none"}, \code{"auto"}, \code{"top"},
\code{"bottom"}. The default is \code{"auto"}.}

\item{caption}{The figure caption.}

\item{label}{Figure label, used for cross-referencing}

\item{supplement}{Whether the figure is to be placed in supplementary material
in the output. This only changes the Typst 'kind' parameter to
\code{"suppl-image"} instead of \code{"image"}. Typst templates may make use of
this to format the figure differently. Default is \code{FALSE.}}

\item{landscape}{Whether the figure should be placed on its own landscape
page. Default is \code{FALSE}.}
}
\value{
A Typst figure with the specified options set.
}
\description{
Set Typst figre options
}
