% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_table_options.R
\name{set_table_options}
\alias{set_table_options}
\title{Set Typst table options}
\usage{
set_table_options(
  x,
  align,
  widths,
  placement,
  caption,
  label,
  footnotes.order,
  footnotes.number,
  footnotes.alphabet,
  footnotes.symbol,
  footnotes.direction,
  supplement,
  landscape,
  na
)
}
\arguments{
\item{x}{A Typst table}

\item{align}{Column alignment. Either \code{"auto"} or a vector with the
same length as the number of columns in \code{x}. Possible values are
\code{"left"},  \code{"center"} (or \code{"centre"}), \code{"right"}. A single non-\code{"auto"}
value will be recycled to the number of columns of \code{x}.}

\item{widths}{Columns widths. Either \code{"auto"} for automatically determined
column widths, \code{"span"} for equal widths spanning the full page, or a
vector of Typst track sizes. A numeric vector will be taken as fractional
lengths.}

\item{placement}{Table placement. Either \code{"none"}, \code{"auto"}, \code{"top"},
\code{"bottom"}. The default is \code{"auto"}.}

\item{caption}{The table caption.}

\item{label}{Table label, used for cross-referencing}

\item{footnotes.order}{The order of different footnote types below the table.
A permumation of \code{c('general', 'number', 'alphabet', 'symbol')}.}

\item{footnotes.number}{The style used for numeric footnote labels. One of
\code{c("arabic", "roman", "Roman")}. Default is \code{"arabic"}.}

\item{footnotes.alphabet}{The style used for alphabetic footnote labels. One
of \code{c("lower", "upper")}. Default is \code{"lower"}.}

\item{footnotes.symbol}{The style used for symbolic footnote labels. One of
\code{c("standard", "extended")} or a character vector of symbols. The
\code{"standard"} set is \code{*}, \verb{†}, \verb{‡}, and \verb{§}; \code{"extended"} adds
\verb{‖} and \verb{¶}. Default is \code{"extended"}.}

\item{footnotes.direction}{The table direction in which footnote numbering is
accumulated. Either \code{"horizontal"} or \code{"vertical"}. Default is
\code{"horizontal"}.}

\item{supplement}{Whether the table is to be placed in supplementary material
in the output. This only changes the Typst 'kind' parameter to
\code{"suppl-table"} instead of \code{"table"}. Typst templates may make use of
this to format the table differently. Default is \code{FALSE.}}

\item{landscape}{Whether the table should be placed on its own landscape
page. Default is \code{FALSE}.}

\item{na}{Character string to print for \code{NA} values. Default is \code{"---"}.}
}
\value{
A Typst table with the specified options set.
}
\description{
Set Typst table options
}
