% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-spans.R
\name{span}
\alias{span}
\alias{colspan}
\alias{rowspan}
\title{Combine table cells spanning multiple columns and/or rows}
\usage{
span(x, start, size, combine = NA_character_)

colspan(x, start, ncols, combine = NA_character_)

rowspan(x, start, nrows, combine = NA_character_)
}
\arguments{
\item{x}{A Typst table}

\item{start}{Top-left cell of the new span. Should be a \code{cells_location}
object created by \code{\link[=cells]{cells()}}.}

\item{size}{A list with elements \code{row} and \code{col} giving the number of rows
and columns the combined cell should span.}

\item{combine}{How the existing contents of the grouped cells should be
combined. Either \code{NA} to keep only the top-left cell or a character
string used to separate the contents of the individual cells. (This may
be revised in future to allow more a flexible specification, using e.g.
\code{glue} syntax.)}

\item{ncols}{The number of columns the combined cell should span}

\item{nrows}{The number of rows the combined cell should span}
}
\value{
A Typst table with the combined cell span added.
}
\description{
Combine table cells spanning multiple columns and/or rows
}
