% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{supplement}
\alias{supplement}
\title{Mark a Typst table as supplementary material}
\usage{
supplement(x)
}
\arguments{
\item{x}{A Typst table}
}
\value{
A Typst table with the supplementary material specification.
}
\description{
Mark a Typst table as supplementary material
}
