% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfig.R
\name{tfig}
\alias{tfig}
\title{Create figures in Typst format}
\usage{
tfig(
  x,
  caption = NULL,
  label = NULL,
  placement = NULL,
  width = NULL,
  height = NULL,
  footnotes = NULL
)
}
\arguments{
\item{x}{Path to a saved image in SVG, PNG, JPEG, or GIF format.}

\item{caption}{The figure caption.}

\item{label}{Figure label, used for cross-referencing}

\item{placement}{(optional) Figure placement. As in Typst's #figure() function.}

\item{width, height}{(optional) Image width and height. As in Typst's #image() function.}

\item{footnotes}{Footnotes to add below the table.}
}
\description{
This wraps the path to a saved image in a Typst figure environment with caption and label.
}
