% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttab.R
\name{ttab}
\alias{ttab}
\title{Create tables in Typst format}
\usage{
ttab(
  x,
  caption = NULL,
  label = NULL,
  rownames = NULL,
  colnames = NULL,
  align = "auto",
  widths = "auto",
  placement = "auto",
  fontsize = NULL,
  gutter = NULL
)
}
\arguments{
\item{x}{An R data frame.}

\item{caption}{The table caption.}

\item{label}{Table label, used for cross-referencing}

\item{rownames}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}. How to treat existing row
names of \code{x}.}

\item{colnames}{Column names to be used in the table. Default is to take the
existing column names of \code{x}.}

\item{align}{Column alignment. Either \code{"auto"} or a vector with the
same length as the number of columns in \code{x}. Each element should be a
Typst \href{https://typst.app/docs/reference/layout/alignment/}{alignment}
specification. A single non-\code{"auto"} value will be recycled to the number
of columns of \code{x}.}

\item{widths}{Columns widths. Either \code{"auto"} for automatically determined
column widths, \code{"span"} for equal widths spanning the full page, or a
vector of Typst \href{https://typst.app/docs/reference/layout/grid/}{track sizes}.
A numeric vector will be taken as fractional lengths.}

\item{placement}{Table placement. Either \code{"none"}, \code{"auto"}, \code{"top"},
\code{"bottom"}. The default is \code{"auto"}. See the Typst
\href{https://typst.app/docs/reference/model/figure/#parameters-placement}{figure}
documentation for details.}

\item{fontsize}{Font size. A Typst
\href{https://typst.app/docs/reference/text/text/#parameters-size}{length}
specification. A numeric \code{fontsize} will be taken as points.}

\item{gutter}{The gaps between table cells. Should be a list with elements
\code{column} and \code{row}, each element being either \code{"auto"} or a Typst
\href{https://typst.app/docs/reference/layout/length/}{length} specification.}
}
\value{
A \code{ttables_tbl} object describing the table content, formatting, and
layout.
}
\description{
This is a table generator inspired by \emph{kableExtra} and \emph{gt}. This base
function takes an R data frame and sets up a basic Typst table. Auxiliary
functions allow the adding of additional formatting features.
}
\examples{
ttab(mtcars)

}
