% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_typst.R
\name{as_typst}
\alias{as_typst}
\title{Print a \code{typst_table} object as raw Typst}
\usage{
as_typst(x)
}
\arguments{
\item{x}{A Typst table}
}
\value{
A \code{glue} string containing the raw Typst output.
}
\description{
Print a \code{typst_table} object as raw Typst
}
